+++ scripts/classes/GmmItem.js
@@ -67,6 +67,17 @@ const GmmItem = (function () {
                 this.system.damage.parts = this.system.damage.parts.map((x) =>
                     (gmmMonster) ? Shortcoder.replaceShortcodesAndAddDamageTypeDamageObject(x[0], gmmMonster, x[1]) : x[0]
                 );
+                this.system._source.description = this.system.description;
+            }
+            return wrapped(...args);
+        }, 'WRAPPER');
+        libWrapper.register('giffyglyph-monster-maker-continued', 'CONFIG.Item.documentClass.prototype.displayCard', function (wrapped, ...args) {
+            if (this.getSheetId() == `${GMM_MODULE_TITLE}.ActionSheet` && this.isOwnedByGmmMonster()) {
+                const gmmMonster = this.getOwningGmmMonster();
+                this.system.damage.parts = this.system.damage.parts.map((x) =>
+                    (gmmMonster) ? Shortcoder.replaceShortcodesAndAddDamageTypeDamageObject(x[0], gmmMonster, x[1]) : x[0]
+                );
+                this.system._source.description = this.system.description;
             }
             return wrapped(...args);
         }, 'WRAPPER');
@@ -240,7 +251,7 @@ const GmmItem = (function () {
                 break;
         }
         let desc = await this.getChatData({ secrets: this.actor?.isOwner });
-        labels.description = desc.description.value;
+        labels.description = Shortcoder.replaceShortcodes(desc.description.value, gmmMonster); ;
 
         if (this.hasLimitedUses) {
             labels.uses = {
@@ -496,7 +507,6 @@ const GmmItem = (function () {
         // Call the roll helper utility
         return damageRoll(mergeObject(rollConfig, options));
     }
-
     function _getSortingCategory() {
         if (this.getSheetId() == `${GMM_MODULE_TITLE}.ActionSheet`) {
             const gmmActionBlueprint = this.getGmmActionBlueprint();