+++ scripts/classes/MonsterSheet.js
@@ -62,6 +62,7 @@ export default class MonsterSheet extends ActorSheet {
 			$el.find('[data-action="delete-item"]').click(this._deleteItem.bind(this));
 			$el.find('[data-action="add-item"]').click(this._addItem.bind(this));
 			$el.find('[data-action="roll-item"]').click(this._rollItem.bind(this));
+			$el.find('[data-action="display-item"]').click(this._displayItem.bind(this));
 			$el.find('[data-action="recharge-item"]').click(this._rechargeItem.bind(this));
 			$el.find('[data-action="update-item"]').change((e) => this._updateItem(e));
 			$el.find('[data-action="roll-hp"]').click((e) => this._rollHitPoints(e));
@@ -207,6 +208,15 @@ export default class MonsterSheet extends ActorSheet {
 		const item = this.actor.items.get(li.dataset.itemId);
 		return item.use();
 	}
+	async _displayItem(event) {
+		const li = event.currentTarget.closest(".item");
+		const item = this.actor.items.get(li.dataset.itemId);
+		const msg = await item.displayCard({ createMessage: false });
+		const DIV = document.createElement("DIV");
+		DIV.innerHTML = msg.content;
+		DIV.querySelector("div.card-buttons").remove();
+		return await ChatMessage.create({ content: DIV.innerHTML });
+	}
 
 	_editItem(event) {
 		const li = event.currentTarget.closest(".item");
@@ -219,7 +229,7 @@ export default class MonsterSheet extends ActorSheet {
 		this.actor.deleteEmbeddedDocuments("Item", [ li.dataset.itemId] );
 	}
 
-	_addItem(event) {
+	async _addItem(event) {
 		const header = event.currentTarget;
 		const type = header.dataset.type;
 		const itemData = {
@@ -232,7 +242,11 @@ export default class MonsterSheet extends ActorSheet {
 				}
 		};
 		delete itemData.system["type"];
-		return this.actor.createEmbeddedDocuments("Item", [ itemData]);
+		delete itemData.system["action"];
+		let item = await this.actor.createEmbeddedDocuments("Item", [itemData]);
+		itemData._id = item[0]._id;
+		this.actor.updateEmbeddedDocuments("Item", [itemData]);
+		return item;
 	}
 
 	_updateAbilityRanking(event) {